pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "automation-hub"
version = "0.1.0"
description = "Centralized automation clients and AI utilities"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # --- Core & Environment ---
    "python-dotenv",
    "requests",
    # --- Google Drive API Integration ---
    "google-api-python-client",
    "google-auth-httplib2",
    "google-auth-oauthlib",
    # --- Data Science & AI (AI-Lab Support) ---
    "numpy>=1.26.0",
    "pandas>=2.2.0",
    "scipy>=1.12.0",
    "matplotlib>=3.8.0",
    "scikit-learn>=1.4.0",
    # --- Quality Gates & Security (SSoT) ---
    "bandit",
    "pip-audit",
    "ruff",
    # --- Testing & Notebook Automation ---
    "pytest",
    "nbclient",
    "nbformat"
]

[tool.setuptools.packages.find]
where = ["."]
include = ["clients"]

# --- UNIFIED QUALITY ENGINE ---
[tool.ruff]
target-version = "py312"
line-length = 88

[tool.ruff.lint]
# E, F: Pyflakes/Error | I: Isort | UP: Upgrade | B: Bugbear | T20: No-Print | S: Security (Bandit)
select = ["E", "F", "I", "UP", "B", "T20", "S", "PT"]
ignore = ["S101"] # Ignore S101 (assert) to allow pytest assertions

[tool.ruff.lint.isort]
combine-as-imports = true
force-single-line = false
known-first-party = ["clients"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.lint.per-file-ignores]
"**/tests/**/*.py" = ["S101"]
"**/*_test.py" = ["S101"]
"**/test_*.py" = ["S101"]
# Allow print only in the core logger client
"clients/core_lib/core_lib_client/logger_client.py" = ["T201"]

[tool.ruff.lint.pycodestyle]
max-doc-length = 100

[tool.pytest.ini_options]
filterwarnings = [
    "ignore::DeprecationWarning:httplib2.*",
    "ignore::DeprecationWarning:googleapiclient.*",
]
addopts = "--doctest-modules --strict-markers"
